+++ dwds/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Expose a `BuildRunnerAssetHandler` which proxies request to the asset server
   running within build runner.
 - Support the Legacy Module strategy through the injected client.
+- Update SDK dependency to minimum of 2.5.0.
 
 ## 0.6.2